feature-policy/experimental-features/focus-without-user-activation-tentative.sub.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./resources/common.js"></script>
<title> 'focus-without-user-activation' Policy : Correctly block automatic focus when policy disabled
</title>
<body>
<script>
  "use strict"

  // Use same origin url here because when iframe document has cross origin
  // url, autofocus will be blocked by default with following console error:
  // "Blocked autofocusing on a form control in a cross-origin subframe."
  const url = "http://{{hosts[alt][www]}}:{{ports[http][0]}}/feature-policy/experimental-features/resources/focus-without-user-activation-iframe-tentative.html";

  function subframe_focused(subframe, event_name) {
    return new Promise(resolve => {
      window.onmessage = m => resolve(m.data.focused);
      subframe.contentWindow.postMessage(event_name, "*");
    });
  }

  promise_test( async (instance) => {
    const frame = createIframe(document.body, {
      sandbox: "allow-scripts allow-same-origin",
      allow: "focus-without-user-activation 'none'",
      src: url
    });

    await wait_for_load(frame);
    assert_false(await subframe_focused(frame, "autofocus"), "'autofocus' should not work.");
    window.focus(); // Reset focus state in subframe.
    assert_false(await subframe_focused(frame, "focus-input"), "'element.focus' should not work.");
    window.focus(); // Reset focus state in subframe.
    assert_false(await subframe_focused(frame, "focus-window"), "'window.focus' should not work.");
    window.focus(); // Reset focus state in subframe.
  }, "When the policy is disabled, 'autofocus' and scripted focus do not focus " +
     "the document.");

  promise_test( async (instance) => {
    const frame = createIframe(document.body, {
      sandbox: "allow-scripts allow-same-origin",
      allow: "focus-without-user-activation *",
      src: url
    });

    await wait_for_load(frame);
    // assert_true(await subframe_focused(frame, "autofocus"), "'autofocus' should work.");
    window.focus(); // Reset focus state in subframe.
    assert_true(await subframe_focused(frame, "focus-input"), "'element.focus' should work.");
    window.focus(); // Reset focus state in subframe.
    assert_true(await subframe_focused(frame, "focus-window"), "'window.focus' should work.");
    window.focus(); // Reset focus state in subframe.
  }, "When the policy is enabled, 'autofocus' and scripted focus do focus " +
     "the document.");
</script>
</body>
